add AttributionMessage
this makes the update functions nicer.

the Msg type is extended with

    | AttributionMsg AttributionMessage

this is a wrapper around messages that are specific to the attribution
part of the application

A nested update is performed in the update function

    -- performs a nested update on the attribution
    let
        ( newAttribution, attributionCommands ) =
            updateAttribution msg model.attribution
    in
        ( { model | attribution = newAttribution }
        , Cmd.map AttributionMsg attributionCommands
        )
We use `Cmd.map` to convert between `Cmd AttributionMessage` and `Cmd
Msg`. The map wraps an AttributionMessage into a AttributionMsg.

Something similar is needed for the view. The function here is
`Html.map`. It converts `Html AttributionMessage` to `Html Msg`.