Refactor Attribution into a separate directory
This removes a lot of code from the main app.

The structure in the directory is the familar one: model msg update view subscriptions.
An imporatant detail is that the Attribution is ignorant of any HTTP
stuff: it doesn't know the details of how communication with the server
works.

This is important: we can work towards one API file, where all
interaction with the server is bundled.

Profiling is not split out yet (and wont be for iteration 3). It is
too small and rudimentary to justify that.